[LCR-237] refactor: ItemInfoResponse DTO 패키지 이동
