+++ src/main/java/com/lgcns/domain/item/repository/ItemRepositoryImpl.java
@@ -3,8 +3,8 @@ package com.lgcns.domain.item.repository;
 import static com.lgcns.domain.item.domain.QItem.item;
 import static com.querydsl.core.types.dsl.Expressions.*;
 
-import com.lgcns.domain.item.client.dto.ItemInfoResponse;
 import com.lgcns.domain.item.client.dto.response.ItemForPaymentResponse;
+import com.lgcns.domain.item.client.dto.response.ItemInfoResponse;
 import com.lgcns.domain.item.dto.response.ItemLocationResponse;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
